Improve connection handling error and refactor connect to a Promise
