+++ src/commands/status.command.ts
@@ -1,15 +1,6 @@
-import { createConnection } from "net";
-
-import { autoStartDaemon } from "./auto-start-deamon";
+import { connect } from "./connect";
 
 export const status = async (names: string[]): Promise<void> => {
-    await autoStartDaemon();
-
-    const client = createConnection(".pm.sock");
-    client.on("connect", () => {
-        client.write(`status ${JSON.stringify(names)}`);
-    });
-    client.on("data", (data) => {
-        console.log(data.toString());
-    });
+    const client = await connect();
+    client.write(`status ${JSON.stringify(names)}`);
 };
